+++ src/app/models/target-components.ts
@@ -442,6 +442,24 @@ const TARGET_GO_FUNCTION_QUERY = gql`
 export const TARGETDETAILSFIELDS = gql`
   fragment targetsDetailsFields on Target {
     ...targetsListFields
+    dataVersions(keys:["Expression", "GTEx"]) {
+      key
+      dataSources {
+        name
+        description
+        url
+        license
+        licenseURL
+        citation
+        files {
+          key
+          file
+          version
+          releaseDate
+          downloadDate
+        }
+      }
+    }
     dataSources
     affiliateLinks: affiliate_links {
       sourceName
@@ -556,31 +574,7 @@ export const TARGETDETAILSFIELDS = gql`
       term
       mimid
     }
-    expressions (top: 10000, filter: {
-      facets: [
-        {
-          facet: "type"
-          values: [
-            "HPA",
-            # "HCA RNA",
-            "JensenLab Experiment HPA",
-            "JensenLab Experiment GNF",
-            "HPM Gene",
-            "HPM Protein",
-            "JensenLab Experiment HPA-RNA",
-            "JensenLab Experiment HPM",
-            "JensenLab Experiment Exon array",
-            "Uniprot Tissue",
-            # "Consensus",
-            "JensenLab Experiment RNA-seq",
-            "JensenLab Knowledge UniProtKB-RC",
-            "JensenLab Text Mining",
-            "JensenLab Experiment UniGene",
-            "JensenLab Experiment Cardiac proteome"
-          ]
-        }
-      ]
-    }) {
+    expressions (top: 10000) {
       type
       sourceRank: source_rank
       tissue
@@ -604,9 +598,12 @@ export const TARGETDETAILSFIELDS = gql`
     }
     gtex {
       tissue
-      gender
       tpm
       tpm_rank
+      tpm_male
+      tpm_male_rank
+      tpm_female
+      tpm_female_rank
       uberon {
         name
         uid