+++ src/app/pharos-main/data-details/target-details/panels/aa-sequence-panel/sequence/sequence.component.ts
@@ -27,7 +27,7 @@ export class SequenceComponent extends DynamicPanelBaseComponent implements OnIn
   }
 
   ngOnInit(): void {
-    if (this.target.sequence) {
+    if (this.target?.sequence) {
       this.parseSequence();
     }
   }