+++ src/app/pharos-main/data-details/target-details/panels/pathways-panel/reactome-pathway-browser/reactome-pathway-browser.component.ts
@@ -1,5 +1,6 @@
 import {Component, Inject, Input, OnChanges, OnInit, PLATFORM_ID, SimpleChanges} from '@angular/core';
-import {isPlatformBrowser} from '@angular/common';
+import {DOCUMENT, isPlatformBrowser} from '@angular/common';
+import {ScriptLoadService} from "../../../../../../pharos-services/script-load.service";
 
 @Component({
   selector: 'pharos-reactome-pathway-browser',
@@ -11,7 +12,8 @@ export class ReactomePathwayBrowserComponent implements OnInit, OnChanges {
   @Input() reactomeID: string;
   @Input() symbol: string;
 
-  constructor(@Inject(PLATFORM_ID) private platformID: any) {
+  constructor(@Inject(PLATFORM_ID) private platformID: any,
+              private scriptLoader: ScriptLoadService) {
   }
 
   fieldChanged(changes: SimpleChanges, fieldName: string) {
@@ -23,24 +25,26 @@ export class ReactomePathwayBrowserComponent implements OnInit, OnChanges {
 
   ngOnChanges(changes: SimpleChanges): void {
     if (this.reactomeID && isPlatformBrowser(this.platformID)) {
-      // @ts-ignore
-      if (this.fieldChanged(changes, 'reactomeID') || this.fieldChanged(changes, 'symbol')) {
-        const that = this;
+      this.scriptLoader.loadReactomeScript().then(() => {
         // @ts-ignore
-        const diagram = Reactome.Diagram.create({
-          placeHolder: 'diagramHolder',
-          width: 900,
-          height: '400'
-        });
-        diagram.loadDiagram(this.reactomeID);
-        diagram.onDiagramLoaded((loaded) => {
-          if (loaded === that.reactomeID) {
-            diagram.resetFlaggedItems();
-            diagram.selectItem(that.reactomeID);
-            diagram.flagItems(that.symbol);
-          }
-        });
-      }
+        if (this.fieldChanged(changes, 'reactomeID') || this.fieldChanged(changes, 'symbol')) {
+          const that = this;
+          // @ts-ignore
+          const diagram = Reactome.Diagram.create({
+            placeHolder: 'diagramHolder',
+            width: 900,
+            height: '400'
+          });
+          diagram.loadDiagram(this.reactomeID);
+          diagram.onDiagramLoaded((loaded) => {
+            if (loaded === that.reactomeID) {
+              diagram.resetFlaggedItems();
+              diagram.selectItem(that.reactomeID);
+              diagram.flagItems(that.symbol);
+            }
+          });
+        }
+      });
     }
   }