+++ test/test-circle-plot.ts
@@ -1,0 +1,10 @@
+export const TestCirclePlot = {
+  name: 'hi',
+  value: 0.25,
+  children: [
+    {name: 'c1',
+    value: 0.5},
+    {name: 'c2',
+    value: 0.75}
+  ]
+}